% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uesd_placebo_plot.R
\name{uesd_placebo_plot}
\alias{uesd_placebo_plot}
\title{Plot the placebo distribution of t‐ or z‐statistics}
\usage{
uesd_placebo_plot(res, binwidth = 0.1)
}
\arguments{
\item{res}{A list returned (invisibly) by \code{uesd_robust_ols(..., placebos = TRUE)}
or \code{uesd_robust_rdd(..., placebos = TRUE)}.}

\item{binwidth}{Numeric. Width of the histogram bins (default = 0.1).}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Given the output from \code{uesd_robust_ols(..., placebos = TRUE)} or
\code{uesd_robust_rdd(..., placebos = TRUE)}, draws a histogram of the
placebo t- or z-values, highlights in red those exceeding the main
statistic, and marks the main value with a vertical line.
}
